backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.5.0
numpy>=1.22.0
Pillow>=9.0
PyYAML>=6.0
scipy>=1.9.0
tqdm>=4.64.0
matplotlib>=3.3
pandas>=1.4.0
seaborn>=0.12.0
requests>=2.28.0
easydict>=1.9
python-dotenv>=1.0.0
